+++ examples/colormaps.py
@@ -21,23 +21,23 @@ plot = mp.vstack(
     mp.text("test images:"),
     mp.hstack(
         mp.border(
-            mp.text("linspace(0,1,256")
-            / mp.image(im_continuous),
+            mp.image(im_continuous),
+            title="linspace(0,1,256",
         ),
         mp.border(
-            mp.text("arange(256)/16")
-            / mp.image(im_discrete16 / 16),
+            mp.image(im_discrete16 / 16),
+            title="arange(256)/16",
         ),
         mp.border(
-            mp.text("arange(256)/25.6")
-            / mp.image(im_discrete10 / 10),
+            mp.image(im_discrete10 / 10),
+            title="arange(256)/25.6",
         )
     ),
     mp.text("continuous colormaps:"),
     mp.wrap(*[
         mp.border(
-            mp.text(c.__name__)
-            / mp.image(im_continuous, colormap=c),
+            mp.image(im_continuous, colormap=c),
+            title=c.__name__,
         )
         for c in [
             mp.reds, mp.greens, mp.blues, mp.rainbow,
@@ -48,14 +48,14 @@ plot = mp.vstack(
     mp.text("discrete colormaps:"),
     mp.wrap(*[
         mp.border(
-            mp.text(c.__name__)
-            / mp.image(im_discrete16, colormap=c),
+            mp.image(im_discrete16, colormap=c),
+            title=c.__name__,
         )
         for c in [ mp.sweetie16, mp.pico8 ]
     ], *[
         mp.border(
-            mp.text(c.__name__)
-            / mp.image(im_discrete10, colormap=c),
+            mp.image(im_discrete10, colormap=c),
+            title=c.__name__,
         )
         for c in [ mp.tableau, mp.nouveau ]
     ], cols=4),